+++ webDeploy.ps1
@@ -22,6 +22,13 @@ This script is used for the backend of the deployment portal
         -Initial Creation
     -Version 2.0 (8/01/2017)
         -Fully Working
+    -Version 3.0 (9/21/2017)
+        -Fixed backup issues
+        -Made backup optional based on leaves
+        -Replaced some True variables with the appropriate $True
+        -Fixed minute issues that had issues with performance
+        -Cleaned and removed unnecessary code.
+        -Early exits on conditions that should cause a termination
 #>
 param(
       [Parameter(Mandatory=$true)]$source,
@@ -101,6 +108,38 @@ else{
    Write-Error -Message ($destComputer +" is not online or is invalid");
    return;
 }
+if(!(Test-Path -Path $source)){
+    Write-Error -Message "Unable to find Source: Check path";
+    return;
+}
+if(!(Test-Path -Path $destination)){
+    echo "The path does not exist, creating folder...";
+    $backup = $false;
+    $folderName = Split-Path $destination -Leaf;
+    $parent = Split-Path $destination -Parent;
+    try{
+        mkdir -Path $parent -Name $folderName;
+    }catch{
+        Write-Error -Message "Unable to create the destination folder... Exiting....";
+        exit;
+    }
+}
+else{
+    $backup = $true;
+    $folderName = Split-Path $destination -Leaf;
+    $parent = Split-Path $destination -Parent;
+    echo "creating a backup for the folder: $foldername";
+    mkdir -Path $parent -name "backup" -ErrorAction Stop;
+    Copy-Item -Path "$destination/*" -Destination "$parent/backup";
+    remove-item -Path "$destination/*" -Force -Recurse -ErrorAction Stop;
+}
+#End confirmation and prep steps.
+
+
+
+
+
+
 
 #Script Blocks
 $scriptContent1 = {
@@ -108,7 +147,6 @@ $scriptContent1 = {
     $source = $args[0];
     $destination = $args[1];
     $appPool = $args[2];
-
     try{
         echo "importing necessary modules on remote server...";
         Import-Module WebAdministration;
@@ -169,50 +207,39 @@ catch
     $results =  "error";
 }
 if(!($results -like "error")){
-
-        #backup goes here
-        try{
-            echo "Creating a backup of the files in destination directory.....";
-            $parent = Split-Path -parent $destination;
-            echo $parent;
-            mkdir -Path $parent -Name "backup" -ErrorAction SilentlyContinue;
-            Copy-Item -Path $destination -Destination "$parent/backup" -Recurse -Force;
-            echo "backup created Successfully!";
-            }
-        catch
-        {
-            Write-Error -Message "Error creating backup....Exiting...";
-            return;
-        }
         try{
             echo "Beginning to copy files...";
             $filesCopied = (Get-ChildItem $source -Recurse | Measure-Object ).Count;
             $values = (Get-ChildItem $source -Recurse | Measure-Object -property length -sum);
             $test = $values.sum / 1MB;
-            $values = ("{0:N2}" -f ($values.sum / 1MB)) + "MB"
-            if($test -gt 250){                                       #Value can be changed
+            $values = ("{0:N2}" -f ($values.sum / 1MB)) + "MB";
+            if($test -gt 250){       #Value can be changed
                 Write-Error -Message "File sizes too large, max of 250MB";
                 return;
             }
             Copy-Item -path $source -Destination $destination -Force -Recurse;
             echo "Files copied successfully: $filesCopied ($values)";
-            echo "Removing backup files...."
-            Remove-Item -Path "$parent/backup" -Recurse -Force;
+            if($backup){
+                echo "Removing backup files...";
+                Remove-Item -Path "$parent/backup" -Recurse -Force;
+            }
         }catch{
-        echo $_.Exception;
-        echo "Restoring backup...."
-        Remove-Item -Path $destination  -Force -Recurse;
-        Copy-Item -Path "$parent/backup" -Destination $destination;
-        Remove-Item -Path "$parent/backup" -Force -Recurse;
-        return;
-    }    
+            if($backup){
+                echo $_.Exception;
+                echo "Restoring backup...."
+                Remove-Item -Path "$destination/*"  -Force -Recurse;
+                Copy-Item -Path "$parent/backup/*" -Destination $destination;
+                Remove-Item -Path "$parent/backup" -Force -Recurse;
+                return;
+            }
+        }    
     Invoke-Command -Session $session -ScriptBlock $scriptContent2 -ArgumentList $appPool;
 }
 else{
     echo "The script terminated as there was no AppPool to stop";
 }
 echo "Script Terminating...";
-Remove-PSSession -Session $session ;
+Remove-PSSession -Session $session;
 return;
